Separate constant files in Kopia package
