+++ pkg/kopia/cmd/const.go
@@ -1,0 +1,98 @@
+package cmd
+
+const (
+	blobSubCommand        = "blob"
+	connectSubCommand     = "connect"
+	createSubCommand      = "create"
+	deleteSubCommand      = "delete"
+	expireSubCommand      = "expire"
+	gcSubCommand          = "gc"
+	infoSubCommand        = "info"
+	kopiaCommand          = "kopia"
+	listSubCommand        = "list"
+	maintenanceSubCommand = "maintenance"
+	manifestSubCommand    = "manifest"
+	policySubCommand      = "policy"
+	repositorySubCommand  = "repository"
+	restoreSubCommand     = "restore"
+	runSubCommand         = "run"
+	setSubCommand         = "set"
+	snapshotSubCommand    = "snapshot"
+	statsSubCommand       = "stats"
+
+	allFlag                    = "--all"
+	bucketFlag                 = "--bucket"
+	cacheDirectoryFlag         = "--cache-directory"
+	configFileFlag             = "--config-file"
+	contentCacheSizeMBFlag     = "--content-cache-size-mb"
+	deleteFlag                 = "--delete"
+	deltaFlag                  = "--delta"
+	endpointFlag               = "--endpoint"
+	filterFlag                 = "--filter"
+	globalFlag                 = "--global"
+	jsonFlag                   = "--json"
+	logDirectoryFlag           = "--log-dir"
+	logLevelErrorFlag          = "--log-level=error"
+	logLevelInfoFlag           = "--log-level=info"
+	metadataCacheSizeMBFlag    = "--metadata-cache-size-mb"
+	noCheckForUpdatesFlag      = "--no-check-for-updates"
+	noGrpcFlag                 = "--no-grpc"
+	noProgressFlag             = "--no-progress"
+	overrideHostnameFlag       = "--override-hostname"
+	overrideUsernameFlag       = "--override-username"
+	parallelFlag               = "--parallel"
+	passwordFlag               = "--password"
+	pointInTimeConnectionFlag  = "--point-in-time"
+	prefixFlag                 = "--prefix"
+	progressUpdateIntervalFlag = "--progress-update-interval"
+	rawFlag                    = "--raw"
+	showIdenticalFlag          = "--show-identical"
+	unsafeIgnoreSourceFlag     = "--unsafe-ignore-source"
+	ownerFlag                  = "--owner"
+	sparseFlag                 = "--sparse"
+
+	// S3 specific
+	s3SubCommand         = "s3"
+	accessKeyFlag        = "--access-key"
+	disableTLSFlag       = "--disable-tls"
+	disableTLSVerifyFlag = "--disable-tls-verification"
+	secretAccessKeyFlag  = "--secret-access-key"
+	sessionTokenFlag     = "--session-token"
+	regionFlag           = "--region"
+
+	// Azure specific
+	azureSubCommand    = "azure"
+	containerFlag      = "--container"
+	storageAccountFlag = "--storage-account"
+	storageKeyFlag     = "--storage-key"
+	storageDomainFlag  = "--storage-domain"
+
+	// Google specific
+	googleSubCommand    = "gcs"
+	credentialsFileFlag = "--credentials-file"
+
+	// Server specific
+	addSubCommand             = "add"
+	refreshSubCommand         = "refresh"
+	serverSubCommand          = "server"
+	startSubCommand           = "start"
+	statusSubCommand          = "status"
+	userSubCommand            = "user"
+	addressFlag               = "--address"
+	redirectToDevNull         = "> /dev/null 2>&1"
+	runInBackground           = "&"
+	serverControlPasswordFlag = "--server-control-password"
+	serverControlUsernameFlag = "--server-control-username"
+	serverPasswordFlag        = "--server-password"
+	serverUsernameFlag        = "--server-username"
+	serverCertFingerprint     = "--server-cert-fingerprint"
+	tlsCertFilePath           = "--tls-cert-file"
+	tlsGenerateCertFlag       = "--tls-generate-cert"
+	tlsKeyFilePath            = "--tls-key-file"
+	urlFlag                   = "--url"
+	userPasswordFlag          = "--user-password"
+
+	// Filesystem specific
+	filesystemSubCommand = "filesystem"
+	pathFlag             = "--path"
+)
